[terra-functional-testing] move code from lib to src in prep of adding the visual-regression-service that needs to be transpiled.
