+++ packages/terra-functional-testing/bin/test-runner.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require('../lib/cli').run();
+require('../src/cli').run();
