+++ packages/terra-functional-testing/src/services/wdio-asset-server-service.js
