+++ packages/terra-functional-testing/src/services/wdio-terra-service.js
