+++ packages/terra-functional-testing/tests/jest/express-server/index.test.js
@@ -1,4 +1,4 @@
-const ExpressServer = require('../../../lib/express-server');
+const ExpressServer = require('../../../src/express-server');
 
 describe('Express Server Index', () => {
   it('should export the express server', () => {
